requirements.txt
python-dotenv==1.0.0
mcphost>=0.1.0
